+++ bonsai_flutter_tool/lib/process_runner.ml
@@ -26,26 +26,67 @@ let status_error program = function
   | Unix.WSTOPPED signal -> Printf.sprintf "%s was stopped by signal %d" program signal
 ;;
 
+let spawn (command : Plan.command) =
+  let environment = merge_environment command.environment in
+  let arguments = Array.of_list (command.program :: command.arguments) in
+  with_working_directory command.working_directory (fun () ->
+    Unix.create_process_env
+      command.program
+      arguments
+      environment
+      Unix.stdin
+      Unix.stdout
+      Unix.stderr)
+;;
+
+let rec waitpid process =
+  try snd (Unix.waitpid [] process) with
+  | Unix.Unix_error (Unix.EINTR, _, _) -> waitpid process
+;;
+
 let run (command : Plan.command) =
   try
-    let environment = merge_environment command.environment in
-    let arguments = Array.of_list (command.program :: command.arguments) in
-    let status =
-      with_working_directory command.working_directory (fun () ->
-        let process =
-          Unix.create_process_env
-            command.program
-            arguments
-            environment
-            Unix.stdin
-            Unix.stdout
-            Unix.stderr
-        in
-        snd (Unix.waitpid [] process))
-    in
+    let status = spawn command |> waitpid in
     match status with
     | Unix.WEXITED 0 -> Ok ()
     | status -> Error (status_error command.program status)
+  with
+  | Unix.Unix_error (error, function_name, argument) ->
+    Error
+      (Printf.sprintf
+         "%s failed for %s: %s"
+         function_name
+         argument
+         (Unix.error_message error))
+;;
+
+let signal_exit_code signal =
+  let number =
+    if signal = Sys.sighup
+    then 1
+    else if signal = Sys.sigint
+    then 2
+    else if signal = Sys.sigquit
+    then 3
+    else if signal = Sys.sigkill
+    then 9
+    else if signal = Sys.sigterm
+    then 15
+    else Int.abs signal
+  in
+  128 + number
+;;
+
+let exit_code = function
+  | Unix.WEXITED code -> code
+  | Unix.WSIGNALED signal | Unix.WSTOPPED signal -> signal_exit_code signal
+;;
+
+let run_status ?(on_spawn = fun _ -> ()) (command : Plan.command) =
+  try
+    let process = spawn command in
+    on_spawn process;
+    Ok (waitpid process |> exit_code)
   with
   | Unix.Unix_error (error, function_name, argument) ->
     Error